requirements.txt
# Core dependencies
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.19.5
opencv-python>=4.5.1
matplotlib>=3.3.4
PyYAML>=5.4.1
tqdm>=4.56.0
pandas>=1.2.1
scikit-learn>=0.24.1
scipy>=1.6.0

# YOLOv7 dependencies (from YOLOv7 repo)
seaborn>=0.11.1
tensorboard>=2.4.1

# Visualization
tensorboardX>=2.1

# Video processing
moviepy>=1.0.3
imageio>=2.9.0

# Utilities
requests>=2.25.1
pillow>=8.1.0